requirements.txt
Flask
gunicorn


pandas
scikit-learn
xgboost

kafka-python
pymongo